metier

+++ mainwindow.cpp
@@ -1,6 +1,7 @@
  #include "mainwindow.h"
 #include "ui_mainwindow.h"
 #include "animaux.h"
+#include "widget.h"
 #include <QIntValidator>
 #include <QMessageBox>
 #include <QSqlQuery>
@@ -31,9 +32,54 @@
 
 #include <QSystemTrayIcon>
 
+#include <QPieSlice>
+#include <QPieSeries>
+#include <QtCharts/QChartView>
+#include <QtWidgets/QApplication>
+#include <QtWidgets/QMainWindow>
+#include <QtCharts/QBarSeries>
+#include <QtCharts/QBarSet>
+#include <QtCharts/QLegend>
+#include <QtCharts>
+#include <QtCharts/QChartView>
+#include <QtCharts/QPieSeries>
+#include <QtCharts/QPieSlice>
 
 
-//#include <QPieSlice>
+
+#include <string>
+#include <vector>
+#include<QDirModel>
+#include <qrcode.h>
+
+#include <iostream>
+#include <fstream>
+#include <QtSvg/QSvgRenderer>
+#include "qrcode.h"
+
+
+
+
+#include<QRegularExpression>
+
+# include "notification.h"
+#include <QPainter>
+
+#include <QSqlQueryModel>
+
+#include <QSystemTrayIcon>
+
+
+#include <QTextStream>
+
+#include<QDirModel>
+
+
+
+
+using qrcodegen::QrCode;
+
+
 
 //test commit men andi
 
@@ -42,23 +88,21 @@ MainWindow::MainWindow(QWidget *parent)
     , ui(new Ui::MainWindow)
 {
     ui->setupUi(this);
-
-        ui->le_idAnim->setValidator( new QIntValidator(100, 99999999, this));
-        ui->le_age->setValidator( new QIntValidator(100, 99, this));
-        ui->tab_animaux->setModel(A.afficherAnim());
+ui->tab_animaux->setModel(A.afficherAnim());
+        ui->le_idAnim->setValidator( new QIntValidator(0, 99999999, this));
+        ui->le_age->setValidator( new QIntValidator(0, 999, this));
         ui->le_nom->setMaxLength(15); // max longueur de nom (line_edit)
         ui->le_status->setMaxLength(15);
+        ui->le_nom->setInputMask("A<AAAAAAAAAAA AAAAAAAA");
+        ui->le_pays->setInputMask("A<AAAAAAAAAAAAAAA");
+        ui->le_rechercher->setPlaceholderText(QString("Research ID ,REGIME , TYPE "));
 
-
-      /*  mysystem = new QSystemTrayIcon(this);
-                             mysystem ->setIcon(QIcon(":/myappico.png"));
-                             mysystem ->setVisible(true);
-*/
 
 }
 MainWindow::~MainWindow()
 {
     delete ui;
+
 }
 
 QModelIndex MainWindow::on_tab_animaux_activated( QModelIndex index)
@@ -68,6 +112,22 @@ QModelIndex MainWindow::on_tab_animaux_activated( QModelIndex index)
 
 }
 
+
+bool MainWindow::search(int t)
+{bool test=false;
+
+       for (int var = 0; var < ui->tab_animaux->model()->rowCount(); ++var) {
+           if(ui->tab_animaux->model()->index(var,0).data().toInt()==t){
+
+              test=true;
+           }
+       }
+       return test;
+}
+
+
+
+/********************************   afficher  *****************************/
 void MainWindow::on_pb_afficher_clicked()
 {
     ui->tab_animaux->setModel(A.afficherAnim());
@@ -75,6 +135,8 @@ void MainWindow::on_pb_afficher_clicked()
 }
 
 
+
+/********************************   ajouter  *****************************/
 void MainWindow::on_bp_ajouter_clicked()
 {
 
@@ -98,9 +160,6 @@ void MainWindow::on_bp_ajouter_clicked()
            {
                ui->tab_animaux->setModel(A.afficherAnim());
                N.notification_ajoutanimau();
-              //  n.setPopupText("Animal a été Ajouté");
-                       // n.show();
-                       //mysystem->showMessage(tr("Notification"),tr("Il y a un Ajout d'un animal"));
 
                 msgBox.setText("Ajout avec succes.");
 
@@ -123,38 +182,46 @@ void MainWindow::on_bp_ajouter_clicked()
 
 
 
-
+/********************************   supprimer  *****************************/
 void MainWindow::on_pb_supprimer_clicked()
 {
 
     Notification N;
+    QMessageBox msgBox;
     QModelIndex index;
     index=on_tab_animaux_activated(index);
     QVariant value=ui->tab_animaux->model()->data(index);
     QString v=value.toString();
 
     ui->le_idAnim->setText(v);
-   /*ui->le_nom->setText(v);
-    ui->le_typeAnimal->setText(v);
-    ui->le_age->setText(v);
-    ui->le_pays->setText(v);
-    ui->comboBox_regime->setCurrentText(v);
-    ui->le_status->setText(v);*/
-
-
 
     animaux A1;
     A1.setidAnim(ui->le_idAnim ->text().toInt());
+
+    int c=A1.getidAnim();
+    bool t=search(c);
+    qDebug() << t;
+
+
+    if(t==false)
+    {
+        QMessageBox::critical(nullptr, QObject::tr("CIN exist"),
+                    QObject::tr("supression failed.\n"
+                                "Click Cancel to exit."), QMessageBox::Cancel);
+    }
+
+    else
+       {
     bool test=A1.supprimerAnim(A1.getidAnim());
 
-    QMessageBox msgBox;
+
 
     if(test)
-       {
-        N.notification_supprimeranimau();
+       { N.notification_supprimeranimau();
         msgBox.setText("suppression  avec succes.");
         ui->tab_animaux->setModel(A1.afficherAnim());
        }
+
     else
         msgBox.setText("Echec de suppression.!!!");
         msgBox.exec();
@@ -167,6 +234,7 @@ void MainWindow::on_pb_supprimer_clicked()
         ui->le_age->clear();
 
 
+     }
 }
 
 /********************************   ACTIVE  TABLE  *****************************/
@@ -187,7 +255,7 @@ void MainWindow::on_tab_animaux_clicked(const QModelIndex &index)
 
 
 
-
+/********************************  modifier  *****************************/
 void MainWindow::on_pb_modifier_clicked()
 {
     Notification N;
@@ -210,9 +278,7 @@ void MainWindow::on_pb_modifier_clicked()
     animaux A(idAnim, nomAnim,typeAnim,pays,ageAnim,status,regimeAliment );
 
 
-
     bool test=A.modifierAnim(A.getidAnim(),A.getNomAnim(),A.gettypeAnim(),A.getpays(),A.getageAnim(),A.getregimealim(),A.getstatus());
-
 
     QMessageBox msgBox;
 
@@ -235,43 +301,56 @@ void MainWindow::on_pb_modifier_clicked()
         ui->le_age->clear();
 }
 
+
+/********************************   rechercher  *****************************/
 void MainWindow::on_pb_rechercher_clicked()
 {
-    animaux A1;
+    /*animaux A1;
     A1.setidAnim(ui->le_idAnim ->text().toInt());
     A1.settypeAnim(ui->le_typeAnimal ->text());
     A1.setregimealim(ui->le_regime->text());
 
         ui->tab_animaux->setModel(A1.chercherAnim(A1.getidAnim(),A1.gettypeAnim(),A1.getregimealim()));
 
-        ui->le_nom->clear();
-        ui->le_pays->clear();
+
+
         ui->le_idAnim->clear();
-        ui->le_status->clear();
         ui->le_typeAnimal->clear();
         ui->le_regime->clear();
-        ui->le_age->clear();
+       */
+
+
+//recherche avancé avec trois critères
+
+    QString rech=ui->le_rechercher->text();
+       ui->tab_animaux->setModel(A.chercheranimal(rech));
+       ui->le_rechercher->clear();
 
 
 
 }
 
 
-void MainWindow::on_pb_trier_clicked()
+/********************************   trier  *****************************/
+
+void MainWindow::on_comboBox_trier_activated()
+{
+    if(ui->comboBox_trier->currentText()=="tri par id")
+        ui->tab_animaux->setModel(A.trierAnim());
+    if(ui->comboBox_trier->currentText()=="tri par age")
+        ui->tab_animaux->setModel(A.trierAnimAge());
+    if(ui->comboBox_trier->currentText()=="tri par pays")
+        ui->tab_animaux->setModel(A.trierAnimpays());
+
+}
+
+
+/*void MainWindow::on_pb_trier_clicked()
 {
 
 
         ui->tab_animaux->setModel(A.trierAnim());
 
-
-       /* int idAnim=ui->le_idAnim->text().toInt();
-        QString nomAnim=ui->le_nom->text();
-        QString typeAnim=ui->le_typeAnimal->text();
-        int ageAnim=ui->spinBox_age->text().toInt();
-        QString pays=ui->le_pays->text();
-        QString regimeAliment=ui->comboBox_regime->currentText();
-        QString status=ui->le_status->text();
-        animaux A(idAnim, nomAnim,typeAnim,pays,ageAnim,status,regimeAliment );*/
 
         QMessageBox msg;
        if(A.trierAnim())
@@ -292,10 +371,10 @@ void MainWindow::on_pb_trier_clicked()
 
 
 
-}
+}*/
 
 
-
+/********************************   reset  *****************************/
 void MainWindow::on_pb_reset_clicked()
 {
     animaux A1;
@@ -324,9 +403,12 @@ void MainWindow::on_pb_reset_clicked()
 
 
 
+/********************************  pdf   *****************************/
 
 void MainWindow::on_pb_pdf_clicked()
 {
+
+
 
     QString strStream;
     strStream = QFileDialog::getSaveFileName((QWidget* )0, "Export PDF", QString(), "*.pdf");
@@ -336,22 +418,24 @@ void MainWindow::on_pb_pdf_clicked()
         printer.setOutputFormat(QPrinter::PdfFormat);
         printer.setPaperSize(QPrinter::A4);
         printer.setOutputFileName(strStream);
-        printer.drawPixmap(QRect(7600,70,2000,2600),QPixmap("C:/Users/Ahmed.A.Hsouna/Desktop/Ahmed_Amin_Ben_Hsouna/logo.png"));
+        //printer.drawPixmap(QRect(7600,70,2000,2600),QPixmap("C:/Users/Ahmed.A.Hsouna/Desktop/Ahmed_Amin_Ben_Hsouna/logo.png"));
 
 
               QTextStream out(&strStream);
 
               const int rowCount = ui->tab_animaux->model()->rowCount();
               const int columnCount = ui->tab_animaux->model()->columnCount();
-              QString TT = QDate::currentDate().toString("yyyy/MM/dd");
+              QString TT = QDateTime::currentDateTime().toString();
               out <<"<html>\n"
                     "<head>\n"
                      "<meta Content=\"Text/html; charset=Windows-1251\">\n"
                   << "<title>ERP - COMmANDE LIST<title>\n "
                   << "</head>\n"
                   "<body bgcolor=#ffffff link=#5000A0>\n"
-                  "<h1 style=\"text-align: center;\"><strong> ******LISTE DES ANIMAUX ****** "+TT+"</strong></h1>"
-                 //   " <img src="logo.png" alt="Trulli" width="500" height="333">"
+                     "<h1 style=\"text-align: center;\"><strong> "+TT+"</strong></h1>"
+                    +"<img src=C:\\Users\\asus\\Desktop\\pdf\\logo_projet />"
+                  "<h1 style=\"text-align: center;\"><strong> ******LISTE DES ANIMAUX ****** </strong></h1>"
+
                   "<table style=\"text-align: center; font-size: 10px;\" border=1>\n "
                     "</br> </br>";
               // headers
@@ -379,6 +463,8 @@ void MainWindow::on_pb_pdf_clicked()
               QTextDocument document;
               document.setHtml(strStream);
               document.print(&printer);
+
+
 
   /*  QPdfWriter pdf("C:\\Users\\asus\\Desktop\\pdf\\pdfanimaux.pdf");
 
@@ -430,7 +516,7 @@ void MainWindow::on_pb_pdf_clicked()
 }
 
 
-
+/********************************  imprimer   *****************************/
 void MainWindow::on_pb_imprimer_clicked()
 {
 
@@ -439,14 +525,18 @@ void MainWindow::on_pb_imprimer_clicked()
 
                  const int rowCount = ui->tab_animaux->model()->rowCount();
                  const int columnCount = ui->tab_animaux->model()->columnCount();
-                 QString TT = QDate::currentDate().toString("yyyy/MM/dd");
+                 QString TT = QDateTime::currentDateTime().toString();
                  out <<"<html>\n"
                        "<head>\n"
                         "<meta Content=\"Text/html; charset=Windows-1251\">\n"
                      << "<title>ERP - COMmANDE LIST<title>\n "
                      << "</head>\n"
                      "<body bgcolor=#ffffff link=#5000A0>\n"
-                     "<h1 style=\"text-align: center;\"><strong> ******LISTE DES  licence commerciale ******"+TT+" </strong></h1>"
+                        "<h1 style=\"text-align: center;\"><strong> "+TT+"</strong></h1>"
+                        +"<img src=C:\\Users\\asus\\Desktop\\pdf\\logo_projet />"
+                        "<h1 style=\"text-align: center;\"><strong> ******LISTE DES  licence commerciale ****** </strong></h1>"
+
+
                      "<table style=\"text-align: center; font-size: 20px;\" border=1>\n "
                        "</br> </br>";
                  // headers
@@ -487,7 +577,162 @@ void MainWindow::on_pb_imprimer_clicked()
 
 
 
-void MainWindow::on_pushButton_clicked()
+
+/********************************  statistiques   *****************************/
+void MainWindow::on_pb_stat_clicked()
+{
+    QSqlQueryModel * model= new QSqlQueryModel();
+                     model->setQuery("select * from animaux where AGE_ANIMAL < 5 ");
+                     float age=model->rowCount();
+                     model->setQuery("select * from animaux where AGE_ANIMAL  between 5 and 10 ");
+                     float agee=model->rowCount();
+                     model->setQuery("select * from animaux where AGE_ANIMAL>10 ");
+                     float ageee=model->rowCount();
+                     float total=age+agee+ageee;
+                     QString a=QString("moins de 5 ans "+QString::number((age*100)/total,'f',2)+"%" );
+                     QString b=QString("entre 5 et 10 Ans"+QString::number((agee*100)/total,'f',2)+"%" );
+                     QString c=QString("+10 Ans"+QString::number((ageee*100)/total,'f',2)+"%" );
+                     QPieSeries *series = new QPieSeries();
+                     series->append(a,age);
+                     series->append(b,agee);
+                     series->append(c,ageee);
+             if (age!=0)
+             {QPieSlice *slice = series->slices().at(0);
+              slice->setLabelVisible();
+              slice->setPen(QPen());}
+             if ( agee!=0)
+             {
+                      // Add label, explode and define brush for 2nd slice
+                      QPieSlice *slice1 = series->slices().at(1);
+                      //slice1->setExploded();
+                      slice1->setLabelVisible();
+             }
+             if(ageee!=0)
+             {
+                      // Add labels to rest of slices
+                      QPieSlice *slice2 = series->slices().at(2);
+                      //slice1->setExploded();
+                      slice2->setLabelVisible();
+             }
+                     // Create the chart widget
+                     QChart *chart = new QChart();
+                     // Add data to chart with title and hide legend
+                     chart->addSeries(series);
+                     chart->setTitle("Pourcentage Par Age :Nombre Des Aimaux "+ QString::number(total));
+                     chart->legend()->hide();
+                     // Used to display the chart
+                     QChartView *chartView = new QChartView(chart);
+                     chartView->setRenderHint(QPainter::Antialiasing);
+                     chartView->resize(1000,500);
+                     chartView->show();
+
+
+             /************************   statistique par regime   *****************************/
+
+
+                     //QSqlQueryModel * model1= new QSqlQueryModel();
+
+                     int i=0,j=0,k=0;
+                     for (int var = 0; var < ui->tab_animaux->model()->rowCount(); ++var) {
+                             if(ui->tab_animaux->model()->index(var,5).data().toString()=="herbivore"){
+
+                                 i++;
+
+                             }
+                             else if(ui->tab_animaux->model()->index(var,5).data().toString()=="carnivore")
+                             {
+                                 j++;
+                             }
+                             else
+                                 k++;
+                         }
+
+
+                                      float total1=i+j+k;
+                                      QString a1=QString("herbivore "+QString::number((i*100)/total1,'f',2)+"%" );
+                                      QString b1=QString("autre"+QString::number((k*100)/total1,'f',2)+"%" );
+                                      QString c1=QString("carnivore"+QString::number((j*100)/total1,'f',2)+"%" );
+                                      QPieSeries *series1 = new QPieSeries();
+                                      series1->append(a1,i);
+                                      series1->append(b1,k);
+                                      series1->append(c1,j);
+                              if (i!=0)
+                              {QPieSlice *slice3 = series1->slices().at(0);
+                               slice3->setLabelVisible();
+                               slice3->setPen(QPen());}
+                              if ( k!=0)
+                              {
+                                       // Add label, explode and define brush for 2nd slice
+                                       QPieSlice *slice4 = series1->slices().at(1);
+                                       //slice4->setExploded();
+                                       slice4->setLabelVisible();
+                              }
+                              if(j!=0)
+                              {
+                                       // Add labels to rest of slices
+                                       QPieSlice *slice5 = series1->slices().at(2);
+                                       //slice1->setExploded();
+                                       slice5->setLabelVisible();
+                              }
+                                      // Create the chart widget
+                                      QChart *chart1 = new QChart();
+                                      // Add data to chart with title and hide legend
+                                      chart1->addSeries(series1);
+                                      chart1->setTitle("Pourcentage Par regime :Nombre Des Aimaux "+ QString::number(total1));
+                                      chart1->legend()->hide();
+                                      // Used to display the chart
+                                      QChartView *chartView1 = new QChartView(chart1);
+                                      chartView1->setRenderHint(QPainter::Antialiasing);
+                                      chartView1->resize(1000,500);
+                                      chartView1->show();
+}
+
+
+
+
+/********************************  qrcode   *****************************/
+void MainWindow::on_pb_code_clicked()
 {
 
+     animaux A;
+    if(ui->tab_animaux->currentIndex().row()==-1)
+           QMessageBox::information(nullptr, QObject::tr("Suppression"),
+                                    QObject::tr("Veuillez Choisir un animal du Tableau.\n"
+                                                "Click Ok to exit."), QMessageBox::Ok);
+       else
+       {
+
+
+
+
+                   A.setidAnim(ui->le_idAnim ->text().toInt());
+                   A.setNomAnim(ui->le_nom ->text());
+                   A.settypeAnim(ui->le_typeAnimal ->text());
+                   A.setageAnim(ui->le_age ->text().toInt());
+                   A.setpays(ui->le_pays ->text());
+                   A.setregimealim(ui->le_regime ->text());
+                   A.setstatus(ui->le_status ->text());
+           // int idAnim=ui->tab_animaux->model()->data(ui->tab_animaux->model()->index(ui->tab_animaux->currentIndex().row(),0)).toInt();
+
+
+
+
+
+         QString  rawQr = "ID_ANIMAL:%1 Nom_ANIMAL:%2 TYPE_ANIMAL:%3 AGE_ANIMAL:%4 PAYS:%5 REGIME_ALIMENT:%6 STATUS:%7 " ;
+            rawQr = rawQr.arg(A.getidAnim()).arg(A.getNomAnim()).arg(A.gettypeAnim()).arg(A.getageAnim()).arg(A.getpays()).arg(A.getregimealim()).arg(A.getstatus());
+            QrCode qr = QrCode::encodeText(rawQr.toUtf8().constData(), QrCode::Ecc::HIGH);
+
+
+          //  const QrCode qr = QrCode::encodeText(std::to_string(idAnim).c_str(), QrCode::Ecc::LOW);
+            std::ofstream myfile;
+            myfile.open ("qrcode.svg") ;
+            myfile << qr.toSvgString(1);
+            myfile.close();
+            QSvgRenderer svgRenderer(QString("qrcode.svg"));
+            QPixmap pix( QSize(140, 140) );
+            QPainter pixPainter( &pix );
+            svgRenderer.render( &pixPainter );
+            ui->label_qrcode->setPixmap(pix);
+       }
 }
+
